requirements_render.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
Authlib==1.6.3
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
coverage==7.10.7
cryptography==44.0.2
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
firebase-admin==6.7.0
Flask==2.3.3
flask-cors==6.0.1
fonttools==4.58.4
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
kiwisolver==1.4.8
# Editable install with no version control (logs-platform==0.1.0)
-e e:\logs_centralizados
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mercadopago==2.3.0
msgpack==1.1.0
narwhals==1.44.0
numpy==2.3.1
openai==2.2.0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.1.0
plotly==6.2.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.11.0
pydantic_core==2.33.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.3
rsa==4.9
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.48.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.29.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
